mutual_funds/app/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
asgiref==3.8.1
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
ecdsa==0.19.0
fastapi==0.115.6
Flask==3.1.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
starlette==0.41.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.34.0
Werkzeug==3.1.3
